refactor(tests): split test_config into test_processors and test_cfstore
